# api/openapi/atlas.openapi.json
{
  "components": {
    "schemas": {
...
      "EntityFlagCreateRequest": {
        "description": "Anonymous flag submission for an entity.",
        "properties": {
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "reason"
        ],
        "title": "EntityFlagCreateRequest",
        "type": "object"
      },
      "EntityFlagListResponse": {
        "description": "Entity flags collection.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FlagResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "EntityFlagListResponse",
        "type": "object"
      },
...
      "FlagResponse": {
        "description": "Flag resource for an entity or source.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target_type",
          "target_id",
          "reason",
          "status",
          "created_at"
        ],
        "title": "FlagResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SourceFlagCreateRequest": {
        "description": "Anonymous flag submission for a source.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "reason"
        ],
        "title": "SourceFlagCreateRequest",
        "type": "object"
      },
      "SourceFlagListResponse": {
        "description": "Source flags collection.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FlagResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "SourceFlagListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "atlas@example.com",
      "name": "Rebuilding America",
      "url": "https://atlas.example.com/"
    },
    "description": "Atlas exposes place-first civic research data through a typed REST API. The public contract includes entities, sources, issue areas, domains, place profiles, issue signals, discovery runs, and anonymous data flags.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Structured civic research data for Atlas places, entities, sources, and discovery runs.",
    "title": "Atlas REST API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/entity-flags": {
      "get": {
        "description": "List anonymous flags that have been submitted for one Atlas entity.",
        "operationId": "listEntityFlags",
        "parameters": [
          {
            "in": "query",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityFlagListResponse"
                }
              }
            },
            "description": "A paginated collection of entity flags."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List entity flags",
        "tags": [
          "flags"
        ]
      },
      "post": {
        "description": "Submit an anonymous flag for an Atlas entity that looks stale or incorrect.",
        "operationId": "createEntityFlag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityFlagCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "The newly created entity flag."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create an entity flag",
        "tags": [
          "flags"
        ]
      }
    },
...
    "/api/source-flags": {
      "get": {
        "description": "List anonymous flags that have been submitted for one Atlas source.",
        "operationId": "listSourceFlags",
        "parameters": [
          {
            "in": "query",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceFlagListResponse"
                }
              }
            },
            "description": "A paginated collection of source flags."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List source flags",
        "tags": [
          "flags"
        ]
      },
      "post": {
        "description": "Submit an anonymous flag for an Atlas source record that looks stale or incorrect.",
        "operationId": "createSourceFlag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceFlagCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            },
            "description": "The newly created source flag."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a source flag",
        "tags": [
          "flags"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://atlas.example.com"
    },
    {
      "url": "http://localhost:8000"
    },
    {
      "url": "/"
    }
  ],
...
}